+++ src/constants-internal.ts
@@ -1,4 +1,17 @@
 /******************************************************************************
+ * VERSIONS
+ */
+
+/**
+ * Current version of @carto/api-client.
+ * @internal
+ */
+export const API_CLIENT_VERSION = __CARTO_API_CLIENT_VERSION;
+
+/** @internal */
+export const V3_MINOR_VERSION = '3.4';
+
+/******************************************************************************
  * DEFAULTS
  */
 
@@ -20,26 +33,33 @@ export const DEFAULT_CLIENT = 'deck-gl-carto';
  */
 export const DEFAULT_GEO_COLUMN = 'geom';
 
-/******************************************************************************
- * ENUMS
+/**
+ * Fastly default limit is 8192; leave some padding.
+ * @internalRemarks Source: @deck.gl/carto
+ * @internal
  */
+export const DEFAULT_MAX_LENGTH_URL = 7000;
 
 /**
+ * @internalRemarks Source: @deck.gl/carto
  * @internal
- * @internalRemarks Source: @carto/constants
  */
-export enum MapType {
-  TABLE = 'table',
-  QUERY = 'query',
-  TILESET = 'tileset',
-}
+export const DEFAULT_TILE_SIZE = 512;
 
 /**
+ * @internalRemarks Source: @deck.gl/carto
  * @internal
- * @internalRemarks Source: @carto/constants
  */
-export enum ApiVersion {
-  V1 = 'v1',
-  V2 = 'v2',
-  V3 = 'v3',
-}
+export const DEFAULT_TILE_RESOLUTION = 0.5;
+
+/**
+ * @internalRemarks Source: @deck.gl/carto
+ * @internal
+ */
+export const DEFAULT_AGGREGATION_RES_LEVEL_H3 = 4;
+
+/**
+ * @internalRemarks Source: @deck.gl/carto
+ * @internal
+ */
+export const DEFAULT_AGGREGATION_RES_LEVEL_QUADBIN = 6;
