+++ src/sources/h3-table-source.ts
@@ -3,7 +3,7 @@
 // Copyright (c) vis.gl contributors
 
 /* eslint-disable camelcase */
-import {DEFAULT_AGGREGATION_RES_LEVEL_H3} from '../constants';
+import {DEFAULT_AGGREGATION_RES_LEVEL_H3} from '../constants-internal';
 import {baseSource} from './base-source';
 import type {
   AggregationOptions,
